% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{dal}
\alias{dal}
\alias{pal}
\alias{qal}
\alias{ral}
\title{The Asymmetric Laplace Distribution}
\usage{
dal(x, mu = 0, sigma = 1, tau = 0.5, log = FALSE)

pal(x, mu = 0, sigma = 1, tau = 0.5)

ral(n, mu = 0, sigma = 1, tau = 0.5)

qal(x, mu = 0, sigma = 1, tau = 0.5)
}
\arguments{
\item{x}{vector of quantiles (\code{dal}, \code{pal}) or probabilities
(\code{qal}).}

\item{mu}{location parameter.}

\item{sigma}{positive scale parameter.}

\item{tau}{skewness parameter (0,1).}

\item{log}{logical; if \code{TRUE}, probabilities are log--transformed.}

\item{n}{number of observations.}
}
\description{
Density, distribution function, quantile function and random generation for
the asymmetric Laplace distribution.
}
\details{
The asymmetric Laplace distribution with parameters (mu, sigma, tau) has
density \deqn{f(x) = \tau(1-\tau)/\sigma e^{-1/(2\sigma) (\theta max(x,0) +
(1 - \theta) max(-x,0))}}
}
\seealso{
\code{\link{lqmm}}, \code{\link{lqm}}
}
\author{
Marco Geraci
}
\keyword{Laplace}
\keyword{asymmetric}
\keyword{distribution}
