% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqmm_buildModelData.R
\name{buildMixedEffectMatrix}
\alias{buildMixedEffectMatrix}
\title{Build the mixed effects model}
\usage{
buildMixedEffectMatrix(modelCall, processedInputData)
}
\arguments{
\item{modelCall}{The [lqmm]-call}

\item{processedInputData}{Output from [buildModelData]}
}
\value{
Generates the mixed effect data and also returns the `contrasts`
 info from [getLqmmContrasts].
}
\description{
Build the mixed effects model
}
