% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqmm_buildModelData.R
\name{buildModelData}
\alias{buildModelData}
\title{Build model data}
\usage{
buildModelData(modelCall, data, extraModelFrameArgs = list(), xlev = NULL)
}
\arguments{
\item{modelCall}{The [lqmm]-call}

\item{data}{The `data.frame` that the object has been called with}

\item{extraModelFrameArgs}{A `list` with the extra arguments to [stats::model.frame]}

\item{xlev}{a named list of character vectors giving the full set of levels
    to be assumed for each factor.}
}
\value{
A list with `dataMix`, `mmf`, `mmf_df`, `group`, `ngroups` and some attributes
 used in the [lqmm] fit.
}
\description{
Helper function for [lqmm] and the predict for allowing custom predictions
}
