% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqmm_S3_coef.R
\name{coef.lqmm}
\alias{coef.lqmm}
\title{Extract LQMM Coefficients}
\usage{
\method{coef}{lqmm}(object, ...)
}
\arguments{
\item{object}{a fitted object of \code{\link{class}} "lqmm".}

\item{\dots}{not used.}
}
\value{
a vector for single quantiles or a matrix for multiple quantiles.
}
\description{
\code{coef} extracts model coefficients from \code{lqmm} objects.
}
\seealso{
\code{\link{lqmm}} \code{\link{summary.lqmm}}
}
\author{
Marco Geraci
}
\keyword{coefficients}
