% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqmm_covHandling.R
\name{covHandling}
\alias{covHandling}
\title{Variance-Covariance Matrix}
\usage{
covHandling(theta, n, cov_name, quad_type)
}
\arguments{
\item{theta}{unique parameters of the variance-covariance matrix of the
random effects as returned by \code{\link{lqmm}} in \code{theta_z}.}

\item{n}{dimension of the vector of random effects.}

\item{cov_name}{see argument \code{covariance} in \code{\link{lqmm}}.}

\item{quad_type}{type of quadrature "c("normal","robust")".}
}
\description{
This is an auxiliary function.
}
\seealso{
\code{\link{VarCorr.lqmm}}
}
\author{
Marco Geraci
}
\keyword{covariance}
