% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{extractBoot}
\alias{extractBoot}
\alias{extractBoot.boot.lqmm}
\title{Extract Fixed and Random Bootstrapped Parameters}
\usage{
extractBoot(object, which = "fixed")
}
\arguments{
\item{object}{an object of \code{\link{class}} \code{boot.lqmm}.}

\item{which}{character indicating whether \code{"fixed"} or \code{"random"}
parameters.}
}
\value{
a matrix of bootstrapped estimates.
}
\description{
This generic function extracts the fixed and random components of
bootstrapped estimates of an \code{lqmm} object.
}
\details{
The \code{"random"} parameters refer to the "raw" parameters of the
variance-covariance matrix of the random effects as returned by
\code{\link{lqmm.fit.gs}} and \code{\link{lqmm.fit.df}}.
}
\examples{


## Orthodont data
data(Orthodont)

# Random intercept model
fit <- lqmm(distance ~ age, random = ~ 1, group = Subject,
	tau = 0.5, data = Orthodont)
fit.boot <- boot(fit)

# extract fixed effects
B <- extractBoot(fit.boot, which = "fixed")

# covariance matrix estimated fixed parameters
cov(B)


}
\seealso{
\code{\link{boot.lqmm}}, \code{\link{lqmm.fit.gs}},
\code{\link{lqmm.fit.df}}
}
\author{
Marco Geraci
}
\keyword{bootstrap}
