% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_gauss.quad.R
\name{gauss.quad.prob}
\alias{gauss.quad.prob}
\title{Gaussian Quadrature}
\source{
Gordon Smyth with contributions from Yifang Hu, Peter Dunn and
Belinda Phipson. (2011). statmod: Statistical Modeling. R package version
1.4.11. \url{https://CRAN.R-project.org/package=statmod}
}
\usage{
gauss.quad.prob(
  n,
  dist = "uniform",
  l = 0,
  u = 1,
  mu = 0,
  sigma = 1,
  alpha = 1,
  beta = 1
)
}
\arguments{
\item{n}{number of nodes and weigh}

\item{dist}{distribution that Gaussian quadrature is based on, one of "uniform", "normal",
"beta" or "gamma"}

\item{l}{lower limit of uniform distribution}

\item{u}{upper limit of uniform distribution}

\item{mu}{mean of normal distribution}

\item{sigma}{standard deviation of normal distribution}

\item{alpha}{positive shape parameter for gamma distribution or first shape parameter for beta
distribution}

\item{beta}{positive scale parameter for gamma distribution or second shape parameter for
beta distribution}
}
\description{
This function calculates nodes and weights for Gaussian quadrature in terms
of probability distributions. See \code{help("gauss.quad.prob")} from
package \code{statmod}.
}
\author{
Original version by Gordon Smyth
}
\keyword{gaussian}
\keyword{quadrature}
