% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqmm_buildModelData.R
\name{getLqmmContrasts}
\alias{getLqmmContrasts}
\title{Get contrast info}
\usage{
getLqmmContrasts(contr, mmf_df)
}
\arguments{
\item{contr}{From the [stats::model.matrix] attribute}

\item{mmf_df}{The model frame from [buildModelData]}
}
\value{
Contrast info
}
\description{
Retrieves the contrast infor
}
