% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqm_S3_logLik.R
\name{logLik.lqm}
\alias{logLik.lqm}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{lqm}(object, ...)
}
\arguments{
\item{object}{an object of \code{\link{class}} "lqm".}

\item{\dots}{not used.}
}
\description{
\code{logLik.lqm} extracts the log-likelihood of a fitted LQM.
}
\seealso{
\code{\link{lqm}} \code{\link{AIC}}
}
\author{
Marco Geraci
}
\keyword{models}
