% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqmm_S3_coef.R, R/lqmm_S3_logLik.R
\name{logLik.lqmm}
\alias{logLik.lqmm}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{lqmm}(object, ...)

\method{logLik}{lqmm}(object, ...)
}
\arguments{
\item{object}{an object of \code{\link{class}} "lqmm".}

\item{\dots}{not used.}
}
\description{
\code{logLik.lqmm} extracts the log-likelihood of a fitted LQMM.

\code{logLik.lqmm} extracts the log-likelihood of a fitted LQMM.
}
\seealso{
\code{\link{lqmm}} \code{\link{AIC}}

\code{\link{lqmm}} \code{\link{AIC}}
}
\author{
Marco Geraci

Marco Geraci
}
\keyword{models}
