% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqmm-package.R
\docType{package}
\name{lqmm-package}
\alias{lqmm-package}
\title{Linear Quantile Models and Linear Quantile Mixed Models

Fit quantile regression models for independent and hierarchical data

\tabular{ll}{ Package: \tab lqmm\cr Type: \tab Package\cr Version: \tab
1.5.5\cr Date: \tab 2019-12-12\cr License: \tab GPL (>=2)\cr LazyLoad: \tab
yes\cr }}
\references{
Geraci M (2014). Linear quantile mixed models: The lqmm package
for Laplace quantile regression. Journal of Statistical Software, 57(13),
1--29. <doi:10.18637/jss.v057.i13>

Geraci M and Bottai M (2007). Quantile regression for longitudinal data
using the asymmetric Laplace distribution. Biostatistics 8(1), 140--154.
<doi:10.1093/biostatistics/kxj039>

Geraci M and Bottai M (2014). Linear quantile mixed models. Statistics and
Computing, 24(3), 461--479. <doi:10.1007/s11222-013-9381-9>.
}
\author{
Marco Geraci

Maintainer: Marco Geraci <geraci@mailbox.sc.edu>
}
\keyword{quantile}
\keyword{regression}
