% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{meanAL}
\alias{meanAL}
\alias{varAL}
\alias{invvarAL}
\title{Functions for Asymmetric Laplace Distribution Parameters}
\usage{
meanAL(mu, sigma, tau)
}
\arguments{
\item{mu}{location parameter.}

\item{sigma}{scale parameter.}

\item{tau}{skewness parameter.}
}
\description{
Accessory functions.
}
\details{
\code{meanAL} computes the mean of an asymmetric Laplace with parameters
\code{mu}, \code{sigma} and \code{tau}.

\code{varAL} computes the variance of an asymmetric Laplace with parameters
\code{sigma} and \code{tau}.

\code{invvarAL} computes the scale parameter of an asymmetric Laplace with
parameter \code{tau} and variance \code{x}.
}
\references{
Yu K and Zhang J (2005). A three-parameter asymmetric Laplace
distribution and its extension. Communications in Statistics-Theory and
Methods 34, 1867--1879.
}
\seealso{
\code{\link{dal}}, \code{\link{mleAL}}
}
\author{
Marco Geraci
}
\keyword{Laplace}
\keyword{asymmetric}
\keyword{distribution}
\keyword{estimation}
\keyword{likelihood}
\keyword{maximum}
