% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqm_S3_predict.R
\name{predict.lqm}
\alias{predict.lqm}
\alias{predict.lqm.counts}
\title{Predictions from LQM Objects}
\usage{
\method{predict}{lqm}(
  object,
  newdata,
  interval = FALSE,
  level = 0.95,
  na.action = na.pass,
  ...
)
}
\arguments{
\item{object}{an lqm or lqm.counts object.}

\item{newdata}{an optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.}

\item{interval}{logical flag. If TRUE, bootstrap percentile intervals for predictions are provided. This argument is for lqm objects only.}

\item{level}{confidence level. This argument is for lqm objects only.}

\item{na.action}{function determining what should be done with missing values in newdata. The default is to predict NA.}

\item{...}{further arguments passed to [boot.lqm].}
}
\value{
a vector or a matrix or an array of predictions.
}
\description{
This function computes predictions based on fitted linear quantile model.
}
\seealso{
\code{\link{residuals.lqm}}, \code{\link{residuals.lqm.counts}},
\code{\link{lqm}}, \code{\link{lqm.counts}}, \code{\link{coef.lqm}},
\code{\link{boot.lqm}}
}
\author{
Marco Geraci
}
\keyword{prediction}
