% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqm_S3_print.R
\name{print.lqm}
\alias{print.lqm}
\alias{print.lqm.counts}
\title{Print LQM Objects}
\usage{
\method{print}{lqm}(x, digits = max(6, getOption("digits")), ...)
}
\arguments{
\item{x}{an \code{lqm} or \code{lqm.counts} object.}

\item{digits}{a non-null value for digits specifies the minimum number of
significant digits to be printed in values.}

\item{\dots}{not used.}
}
\description{
Print an object generated by \code{\link{lqm}} or \code{\link{lqm.counts}}.
}
\seealso{
\code{\link{lqm}}, \code{\link{lqm.counts}}
}
\author{
Marco Geraci
}
\keyword{print}
