% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqmm_S3_print.R
\name{print.lqmm}
\alias{print.lqmm}
\title{Print an \code{lqmm} Object}
\usage{
\method{print}{lqmm}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an \code{lqmm} object.}

\item{digits}{a non-null value for digits specifies the minimum number of
significant digits to be printed in values.}

\item{\dots}{not used.}
}
\description{
Print an object generated by \code{\link{lqmm}}.
}
\seealso{
\code{\link{lqmm}}
}
\author{
Marco Geraci
}
\keyword{print}
