% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqm_S3_summary.R
\name{print.summary.lqm}
\alias{print.summary.lqm}
\title{Print an \code{lqm} Summary Object}
\usage{
\method{print}{summary.lqm}(x, ...)
}
\arguments{
\item{x}{a \code{summary.lqm} object.}

\item{\dots}{not used.}
}
\description{
Print summary of an \code{lqm} object.
}
\seealso{
\code{\link{lqm}}, \code{\link{summary.lqm}}
}
\author{
Marco Geraci
}
\keyword{print}
\keyword{summary}
