% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqmm_S3_coef.R, R/lqmm_S3_summary.R
\name{print.summary.lqmm}
\alias{print.summary.lqmm}
\title{Print an \code{lqmm} Summary Object}
\usage{
\method{print}{summary.lqmm}(x, digits = max(3, getOption("digits") - 3), ...)

\method{print}{summary.lqmm}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{a \code{summary.lqmm} object.}

\item{digits}{a non-null value for digits specifies the minimum number of
significant digits to be printed in values.}

\item{\dots}{not used.}
}
\description{
Print summary of an \code{lqmm} object.

Print summary of an \code{lqmm} object.
}
\seealso{
\code{\link{lqmm}}, \code{\link{summary.lqmm}}

\code{\link{lqmm}}, \code{\link{summary.lqmm}}
}
\author{
Marco Geraci

Marco Geraci
}
\keyword{print}
\keyword{summary}
