% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqmm_S3_coef.R, R/lqmm_S3_ranef.R
\name{ranef.lqmm}
\alias{ranef.lqmm}
\alias{ranef}
\title{Extract Random Effects}
\usage{
\method{ranef}{lqmm}(object, ...)

\method{ranef}{lqmm}(object, ...)
}
\arguments{
\item{object}{an object of \code{\link{class}} \code{lqmm}.}

\item{\dots}{not used.}
}
\value{
a data frame or a list of data frames of predicted random effects.

a data frame or a list of data frames of predicted random effects.
}
\description{
This function computes random effects for a linear quantile mixed model.

This function computes random effects for a linear quantile mixed model.
}
\details{
The prediction of the random effects is done via estimated best linear
prediction (Geraci and Bottai, 2014). The generic function \code{ranef} is
imported from the \code{nlme} package (Pinheiro et al, 2014).

The prediction of the random effects is done via estimated best linear
prediction (Geraci and Bottai, 2014). The generic function \code{ranef} is
imported from the \code{nlme} package (Pinheiro et al, 2014).
}
\references{
Geraci M and Bottai M (2014). Linear quantile mixed models.
Statistics and Computing, 24(3), 461--479. doi: 10.1007/s11222-013-9381-9.

Pinheiro J, Bates D, DebRoy S, Sarkar D and R Core Team (2014). nlme: Linear
and Nonlinear Mixed Effects Models. R package version 3.1-117,
\url{https://CRAN.R-project.org/package=nlme}.

Geraci M and Bottai M (2014). Linear quantile mixed models.
Statistics and Computing, 24(3), 461--479. doi: 10.1007/s11222-013-9381-9.

Pinheiro J, Bates D, DebRoy S, Sarkar D and R Core Team (2014). nlme: Linear
and Nonlinear Mixed Effects Models. R package version 3.1-117,
\url{https://CRAN.R-project.org/package=nlme}.
}
\seealso{
\code{\link{lqmm}}, \code{\link{coef.lqmm}}

\code{\link{lqmm}}, \code{\link{coef.lqmm}}
}
\author{
Marco Geraci

Marco Geraci
}
\keyword{coefficients}
\keyword{effects}
\keyword{random}
