% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqm_S3_residuals.R
\name{residuals.lqm}
\alias{residuals.lqm}
\alias{residuals.lqm.counts}
\title{Residuals from an LQM Objects}
\usage{
\method{residuals}{lqm}(object, ...)
}
\arguments{
\item{object}{an \code{lqm} or \code{lqm.counts} object.}

\item{\dots}{not used.}
}
\value{
a vector or matrix of residuals.
}
\description{
This function computes the residuals from a fitted linear quantile model.
}
\seealso{
\code{\link{lqm}}, \code{\link{lqm.counts}},
\code{\link{predict.lqm}}, \code{\link{coef.lqm}}
}
\author{
Marco Geraci
}
\keyword{residuals}
