% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{retrieveExpression}
\alias{retrieveExpression}
\title{Retrieves an expression from parent environments}
\usage{
retrieveExpression(expr)
}
\arguments{
\item{expr}{The expression to search for}
}
\value{
The evaluated object
}
\description{
As we update models objects may end up in different environments
and we need to be able to locate them regardless of where they exist.
This function traverses through all parent environments in search for
the object of interest.
}
