% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_gauss.quad.R
\name{gauss.quad.prob}
\alias{gauss.quad.prob}
\title{Gaussian Quadrature}
\source{
Gordon Smyth with contributions from Yifang Hu, Peter Dunn and
Belinda Phipson. (2011). statmod: Statistical Modeling. R package version
1.4.11. \url{https://CRAN.R-project.org/package=statmod}
}
\usage{
gauss.quad.prob(
  n,
  dist = "uniform",
  l = 0,
  u = 1,
  mu = 0,
  sigma = 1,
  alpha = 1,
  beta = 1
)
}
\description{
This function calculates nodes and weights for Gaussian quadrature in terms
of probability distributions. See \code{help("gauss.quad.prob")} from
package \code{statmod}.
}
\author{
Original version by Gordon Smyth
}
\keyword{gaussian}
\keyword{quadrature}
