% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_gauss.quad.R
\name{gauss.quad}
\alias{gauss.quad}
\title{Gaussian Quadrature}
\source{
Gordon Smyth with contributions from Yifang Hu, Peter Dunn and
Belinda Phipson. (2011). statmod: Statistical Modeling. R package version
1.4.11. \url{https://CRAN.R-project.org/package=statmod}
}
\usage{
gauss.quad(n, kind = "legendre", alpha = 0, beta = 0)
}
\arguments{
\item{n}{number of nodes and weigh}

\item{kind}{kind of Gaussian quadrature, one of "legendre", "chebyshev1", "chebyshev2",
"hermite", "jacobi" or "laguerre"}

\item{alpha}{parameter for Jacobi or Laguerre quadrature, must be greater than -1}

\item{beta}{parameter for Jacobi quadrature, must be greater than -1}
}
\description{
This function calculates nodes and weights for Gaussian quadrature. See
\code{help("gauss.quad")} from package \code{statmod}.
}
\author{
Original version by Gordon Smyth
}
\keyword{gaussian}
\keyword{quadrature}
