% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_positive.definite.R
\name{is.positive.definite}
\alias{is.positive.definite}
\title{Test for Positive Definiteness}
\source{
Juliane Schaefer, Rainer Opgen-Rhein, Verena Zuber, A. Pedro Duarte
Silva and Korbinian Strimmer. (2011). corpcor: Efficient Estimation of
Covariance and (Partial) Correlation. R package version 1.6.0.
\url{https://CRAN.R-project.org/package=corpcor}
}
\usage{
is.positive.definite(m, tol, method = c("eigen", "chol"))
}
\description{
This function tests whether all eigenvalues of a symmetric matrix are
positive. See \code{help("is.positive.definite")} from package
\code{corpcor}.
}
\author{
Original version by Korbinian Strimmer
}
\keyword{covariance}
\keyword{definite}
\keyword{positive}
