% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_positive.definite.R
\name{make.positive.definite}
\alias{make.positive.definite}
\title{Compute Nearest Positive Definite Matrix}
\source{
Juliane Schaefer, Rainer Opgen-Rhein, Verena Zuber, A. Pedro Duarte
Silva and Korbinian Strimmer. (2011). corpcor: Efficient Estimation of
Covariance and (Partial) Correlation. R package version 1.6.0.
\url{https://CRAN.R-project.org/package=corpcor}
}
\usage{
make.positive.definite(m, tol)
}
\description{
This function computes the nearest positive definite of a real symmetric
matrix. See \code{help("make.positive.definite")} from package
\code{corpcor}.
}
\author{
Original version by Korbinian Strimmer
}
\keyword{covariance}
\keyword{definite}
\keyword{positive}
